+++ lib/main.dart
@@ -1,13 +1,11 @@
-import 'package:cafe_app/home_screen.dart';
-import 'package:cafe_app/model/settings.dart';
+import 'package:cafe_app/my_app.dart';
 import 'package:cafe_app/utils/custom_theme.dart';
 import 'package:cafe_app/utils/data_provider/theme_provider.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 void main() {
-  runApp(const MyAppWrapper());
+  runApp(const ProviderScope(child: MyAppWrapper()));
 }
 
 class MyAppWrapper extends StatelessWidget {
@@ -18,33 +16,6 @@ class MyAppWrapper extends StatelessWidget {
     return const CustomTheme(
       initialThemeKey: MyThemeKeys.light,
       child: MyApp(),
-    );
-  }
-}
-
-class MyApp extends StatefulWidget {
-  const MyApp({super.key});
-
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (context) => SettingsModel(),
-      child: Consumer<SettingsModel>(
-        builder: (context, settingsModel, child) => MaterialApp(
-          title: 'Cafe app',
-          debugShowCheckedModeBanner: false,
-          theme: CustomTheme.of(context),
-          localizationsDelegates: AppLocalizations.localizationsDelegates,
-          supportedLocales: AppLocalizations.supportedLocales,
-          locale: settingsModel.locale,
-          home: const HomeNavigation(),
-        ),
-      ),
     );
   }
 }
